+++ src/collective/contact/plonegroup/browser/settings.py
@@ -144,7 +144,7 @@ def addOrModifyGroup(group_name, organization_title, function_title):
         )
         return True
     else:
-        #group_title is maybe modified
+        # group_title is maybe modified
         if group.getProperty('title') != group_title:
             group.setProperties(title=group_title)
             return True
@@ -365,6 +365,6 @@ class SelectedOrganizationsElephantVocabulary(object):
                 del terms[uid]
         extra_uids = terms.keys()
         extra_terms = terms.values()
-        ordered_vocab = SimpleVocabulary(ordered_terms+extra_terms)
+        ordered_vocab = SimpleVocabulary(ordered_terms + extra_terms)
         wrapped_vocab = wrap_vocabulary(ordered_vocab, hidden_terms=extra_uids)(context)
         return wrapped_vocab