+++ src/collective/contact/plonegroup/config.py
@@ -1,4 +1,3 @@
-
 # Registry keys
 ORGANIZATIONS_REGISTRY = 'collective.contact.plonegroup.browser.settings.IContactPlonegroupConfig.organizations'
 FUNCTIONS_REGISTRY = 'collective.contact.plonegroup.browser.settings.IContactPlonegroupConfig.functions'