+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-0.3 (unreleased)
+1.0 (2015-11-24)
 ----------------
 
 - Added link integrity check when deleting a plonegroup organization